+++ src/components/Home/Tags/Tags.module.css
@@ -1,0 +1,40 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    width: 268px;
+    height: 124px;
+    background: #1C1C21;
+    border-radius: 20px;
+    padding: 16px;
+}
+
+.title {
+    font-weight: 600;
+    font-size: 16px;
+    margin: 0;
+}
+
+.tags {
+    width: 100%;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: flex-end;
+    gap: 4px;
+    margin-top: 16px;
+}
+
+.tag {
+    font-weight: 600;
+    font-size: 12px;
+    line-height: 16px;
+    color: #fff;
+    padding: 4px 8px;
+    background: #2F2F37;
+    border-radius: 100px;
+}
+
+.tag:focus, .tag:hover {
+    text-decoration: none;
+    background: #4C4CFF;
+    color: #fff;
+}
